anomaly/requirements.txt
fastapi[standard]==0.115.0
pydantic-settings==2.7.0
structlog==25.1.0
scikit-learn==1.6.0
numpy==2.2.0
joblib==1.4.0
httpx==0.28.0

pytest==8.3.0
pytest-asyncio==0.24.0
ruff==0.8.0
mypy==1.14.0
